requirements.txt
Flask
Flask-SocketIO
pymongo
pycryptodome
python-dotenv